Move AtomInfo to separate file
